Add Ng2FlatpickrDirective
Allows users to define an HTML element with a [flatpickr] selector to invoke Flatpickr. They can pass the config object into the [flatpickr] selector (i.e., if it’s defined in the component class), they can pass in any number of config options as attributes on the HTML element, or they can use both — the attribute will override the “global” configuration.